requirements.txt
matplotlib==3.7.2
numpy==1.24.1
open3d==0.17.0
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
Pillow==11.0.0
plyfile==1.1
Requests==2.32.3
scikit_learn==1.3.2
scipy==1.14.1
tensorboardX==2.6.2.2
torch==2.0.0+cu118
torchvision==0.15.0+cu118
tqdm==4.65.2
